Added LMS Candidate Service

+++ src/main/java/com/bridgelabz/lmscandidateservice/service/CandidateService.java
@@ -64,58 +64,70 @@ public class CandidateService implements ICandidateService{
 
     @Override
     public Response updateCandidate(long id, CandidateDTO candidateDTO, String token) {
-        Long candidateId = tokenUtil.decodeToken(token);
-        Optional<CandidateModel> isCandidatePresent = candidateRepository.findById(candidateId);
-        if (isCandidatePresent.isPresent()) {
-            isCandidatePresent.get().setCicId(candidateDTO.getCicId());
-            isCandidatePresent.get().setFullName(candidateDTO.getFullName());
-            isCandidatePresent.get().setMobileNum(candidateDTO.getMobileNum());
-            isCandidatePresent.get().setHiredData(candidateDTO.getHiredData());
-            isCandidatePresent.get().setDegree(candidateDTO.getDegree());
-            isCandidatePresent.get().setAggrPer(candidateDTO.getAggrPer());
-            isCandidatePresent.get().setCity(candidateDTO.getCity());
-            isCandidatePresent.get().setState(candidateDTO.getState());
-            isCandidatePresent.get().setPreferredJobLocation(candidateDTO.getPreferredJobLocation());
-            isCandidatePresent.get().setStatus(candidateDTO.getStatus());
-            isCandidatePresent.get().setCreationTimeStamp(candidateDTO.getCreationTimeStamp());
-            isCandidatePresent.get().setUpdatedTimeStamp(candidateDTO.getUpdatedTimeStamp());
-            String body = "Candidate updated Successfully with id  :" + isCandidatePresent.get().getId();
-            String subject = "Candidate updated Successfully....";
-            mailService.send(isCandidatePresent.get().getEmail(), body, subject);
-            return new Response("Candidate Added Successfully", 200, isCandidatePresent);
+        boolean isUserPresent = restTemplate.getForObject("http://localhost:8080/admin/validate/" + token, Boolean.class);
+        if (isUserPresent) {
+            Long candidateId = tokenUtil.decodeToken(token);
+            Optional<CandidateModel> isCandidatePresent = candidateRepository.findById(candidateId);
+            if (isCandidatePresent.isPresent()) {
+                isCandidatePresent.get().setCicId(candidateDTO.getCicId());
+                isCandidatePresent.get().setFullName(candidateDTO.getFullName());
+                isCandidatePresent.get().setMobileNum(candidateDTO.getMobileNum());
+                isCandidatePresent.get().setHiredData(candidateDTO.getHiredData());
+                isCandidatePresent.get().setDegree(candidateDTO.getDegree());
+                isCandidatePresent.get().setAggrPer(candidateDTO.getAggrPer());
+                isCandidatePresent.get().setCity(candidateDTO.getCity());
+                isCandidatePresent.get().setState(candidateDTO.getState());
+                isCandidatePresent.get().setPreferredJobLocation(candidateDTO.getPreferredJobLocation());
+                isCandidatePresent.get().setStatus(candidateDTO.getStatus());
+                isCandidatePresent.get().setCreationTimeStamp(candidateDTO.getCreationTimeStamp());
+                isCandidatePresent.get().setUpdatedTimeStamp(candidateDTO.getUpdatedTimeStamp());
+                String body = "Candidate updated Successfully with id  :" + isCandidatePresent.get().getId();
+                String subject = "Candidate updated Successfully....";
+                mailService.send(isCandidatePresent.get().getEmail(), body, subject);
+                return new Response("Candidate Added Successfully", 200, isCandidatePresent);
+            }
         }
         throw  new AdminException(400, "Admin not Present");
     }
 
     @Override
      public Response deleteCandidate(Long id, String token) {
-        Long candidateId = tokenUtil.decodeToken(token);
-        Optional<CandidateModel> isCandidatePresent = candidateRepository.findById(candidateId);
-        if (isCandidatePresent.isPresent()){
-            candidateRepository.delete(isCandidatePresent.get());
-            return new Response("Candidate Added Successfully", 200, isCandidatePresent);
+        boolean isUserPresent = restTemplate.getForObject("http://localhost:8080/admin/validate/" + token, Boolean.class);
+        if (isUserPresent) {
+            Long candidateId = tokenUtil.decodeToken(token);
+            Optional<CandidateModel> isCandidatePresent = candidateRepository.findById(candidateId);
+            if (isCandidatePresent.isPresent()) {
+                candidateRepository.delete(isCandidatePresent.get());
+                return new Response("Candidate Added Successfully", 200, isCandidatePresent);
+            }
         }
         throw new AdminException(400, "Candidate Not found");
     }
 
     @Override
     public Response getCandidateStatus(String token, String status) {
-        Long candidateId = tokenUtil.decodeToken(token);
-        List<CandidateModel> isStatusPresent = candidateRepository.findByStatus(status);
-        if (isStatusPresent.size()>0){
-            return new Response("Candidate Added Successfully", 200, isStatusPresent);
+        boolean isUserPresent = restTemplate.getForObject("http://localhost:8080/admin/validate/" + token, Boolean.class);
+        if (isUserPresent) {
+            Long candidateId = tokenUtil.decodeToken(token);
+            List<CandidateModel> isStatusPresent = candidateRepository.findByStatus(status);
+            if (isStatusPresent.size() > 0) {
+                return new Response("Candidate Added Successfully", 200, isStatusPresent);
+            }
         }
         throw new AdminException(400, "Status not found");
     }
 
     @Override
     public Response changeCandidateStatus(long id, String status, String token) {
-        Long candidateId = tokenUtil.decodeToken(token);
-        Optional<CandidateModel> isIdPresent = candidateRepository.findById(candidateId);
-        if (isIdPresent.isPresent()){
-            isIdPresent.get().setStatus(status);
-            candidateRepository.save(isIdPresent.get());
-            return new Response("Candidate Added Successfully", 200, isIdPresent);
+        boolean isUserPresent = restTemplate.getForObject("http://localhost:8080/admin/validate/" + token, Boolean.class);
+        if (isUserPresent) {
+            Long candidateId = tokenUtil.decodeToken(token);
+            Optional<CandidateModel> isIdPresent = candidateRepository.findById(candidateId);
+            if (isIdPresent.isPresent()) {
+                isIdPresent.get().setStatus(status);
+                candidateRepository.save(isIdPresent.get());
+                return new Response("Candidate Added Successfully", 200, isIdPresent);
+            }
         }
         throw new AdminException(400, "Status Not Found");
     }
@@ -123,9 +135,9 @@ public class CandidateService implements ICandidateService{
     @Override
     public Response getStatusCount(String status) {
             List<CandidateModel> isStatusPresent = candidateRepository.findByStatus(status);
-            if (isStatusPresent.size()>0){
+            if (isStatusPresent.size() > 0) {
                 return new Response("Candidate Added Successfully", 200, isStatusPresent);
-            }
+        }
             throw new AdminException(400, "No Status Available");
     }
 }
